+++ cs/FileInfo.cs
@@ -1,4 +1,4 @@
-/*! cs/FileInfo.cs | MIT License | github.com/visicode/Balzac.cs */
+/*! ./cs/FileInfo.cs | MIT License | github.com/visicode/Balzac.cs */
 using System.Diagnostics;
 using System.IO;
 
@@ -17,7 +17,7 @@ namespace Balzac.cs {
 		/// <summary>Tries to delete a file.</summary>
 		/// <returns>true if the file has been deleted, otherwise false.</returns>
 		public static bool TryDelete(this FileInfo fileInfo) {
-			if (fileInfo != null && fileInfo.Exists)
+			if (fileInfo.Exists)
 				try {
 					fileInfo.Delete();
 #if DEBUG