+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changes to Balzac.cs
 
+### 0.1.4 (September 6, 2023)
+
+* Improve String.TrimLines method.
+* Improve WebUtility.ToPlainText method.
+* Add Enum.GetDescription function.
+* Move Test.cs and Test.csproj files to root folder.
+
+### 0.1.3 (September 5, 2023)
+
+* Rename HtmlHelper class to WebUtility.
+* Move String.StripSpecialChars, String.Nl2br, String.Nl2p, String.ToFileName and String.ToPlainText methods to WebUtility class.
+* Add mappings to call initial WebUtility class methods.
+* Remove nullable parameters and returns where possible.
+* Update README.md file.
+
 ### 0.1.2 (August 13, 2023)
 
 * Initial release, based on Balzac.js JavaScript library current version number.